backend-multiple/requirements.txt
Flask
pyrebase4
requests
python-dotenv

